# Use memcpy and memset from musl for performance benefit