--- a/risc0/zkvm/sdk/rust/guest/src/lib.rs
+++ b/risc0/zkvm/sdk/rust/guest/src/lib.rs
@@ -34,7 +34,12 @@ pub mod sha;
 #[cfg(feature = "pure-prove")]
 pub mod sha_insecure;
 
-use core::{arch::asm, mem, ptr};
+use core::{arch::asm, arch::global_asm, mem, ptr};
+
+#[cfg(target_arch = "riscv32")]
+global_asm!(include_str!("memset.s"));
+#[cfg(target_arch = "riscv32")]
+global_asm!(include_str!("memcpy.s"));
 
 extern "C" {
     fn _fault() -> !;