@jbruestle 2022-09-14
CI is angry, looks like you need to run rustfmt.