@stskeeps 2022-09-14
> CI is angry, looks like you need to run rustfmt.

Indeed, it had the right to be annoyed, pushed an update.