@stskeeps 2022-09-14
That error looks more like some kind of CI/build system/cache error, the commit clearly includes risc0/zkvm/sdk/rust/guest/src/memcpy.s as a path and https://doc.rust-lang.org/std/macro.include_str.html says it's located relative to the current file..