@stskeeps 2022-09-14
Fixed, https://github.com/bazelbuild/rules_rust/issues/459