@flaub 2022-09-14
Can you include the copyright header in each `.s`?