@stskeeps 2022-09-15
Fixes:
- Seems like the guest sdk was built for x86_64 in bazel too, made the global_asm only active on riscv32
- Added copyright from musl to the assembly files